backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain
langchain==0.1.0
langchain-community==0.0.10

# Vector database
chromadb==0.4.18

# Document processing
pypdf==3.17.1
python-docx==1.1.0

# HTTP requests
requests==2.31.0

# Data handling
numpy==1.24.3
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0